add getHomeState(), setAway() and setHome()
i got this up and running with an older version of pyTado, but this should work, though i didn’t test it. 
to integrate this in homeassistant i made a sensor from the getHomeState() return value and derived a binary sensor from the showHomePresenceSwitchButton attribute.
as soon as showHomePresenceSwitchButton is not none anymore an automation toogles the home state. this mimics the auto assist skill pretty well.
also: i don’t know if or how any of this works if the auto assist skill is active.