Add tests and check for auto geofencing support
Add check to ensure auto geofencing mode can only be enabled if it is known to be supported, i.e. ensure no issues occur for users without the Auto Assist skill.  Add a custom exception to handle the event something tries to set auto mode when it is not known to be supported
Add autoGeofencingSupported flag to track this ability and an external function for getting it
Add tests